[#27924] docdb: Fix signed overflow issue for max_length for FormatBytesAsStr
Summary:
`FormatBytesAsStr` converts `n` (length) from unsigned to signed integer. But the check
for `max_length` is `max_length > 1`, which causes the loop to not exit properly when `n` is
sufficiently large and the cast to signed integer overflows:

```
  int64_t bytes_left = static_cast<int64_t>(n - 1);
  for (const char* p = data; p != end; ++p, --bytes_left) {
    // ...

    if (result.size() >= max_length && bytes_left > 1) {
      result.append(Substitute("<...$0 bytes skipped>", bytes_left));
      break;
    }
  }
```

This fixes the issue by keeping `bytes_left` as unsigned. In the case that `n == 0` (so `n - 1`
wraps around), `data == end` so we don't enter the loop at all, so there is no need to cast
to signed integer.
Jira: DB-17540

Test Plan:
Added tests:
`./yb_build.sh --cxx_test 'util_bytes_formatter-test' --gtest_filter 'BytesFormatterTest.TestEmpty'`
`./yb_build.sh --cxx_test 'util_bytes_formatter-test' --gtest_filter 'BytesFormatterTest.TestOverflow'`

Reviewers: sergei

Reviewed By: sergei

Subscribers: ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D45331

+++ src/yb/util/bytes_formatter.cc
@@ -44,8 +44,9 @@ string FormatBytesAsStr(const char* data,
   result.push_back(quote);
   const char* end = data + n;
 
-  // Not including the current character we're looking at. Cast to a signed int to avoid underflow.
-  int64_t bytes_left = static_cast<int64_t>(n - 1);
+  // Not including the current character we're looking at.
+  // In case of n == 0, data == end, so we don't enter loop at all.
+  size_t bytes_left = n - 1;
 
   for (const char* p = data; p != end; ++p, --bytes_left) {
     uint8_t c = static_cast<uint8_t>(*p);
